refactor(marketing): move buyer landings + pricing + legal under /marketing/* prefix
CHAOS-1647 shipped four buyer landing pages but planted them at the URL root
(/vp-engineering, /platform-devex, etc.) using the (marketing) route group
purely to share a layout. This was misuse of the route-group mechanism:
parens-folders are for layout sharing within a section, not for hiding an
entire content surface at the root. The result was that /pricing, /privacy,
/terms, and four buyer landings all sat at the same URL depth as /dashboard
and /work, with no URL-level distinction between marketing and app content.
The buyer pages also had zero inbound links from nav or footer, leaving them
orphaned and unreachable except by direct URL.

This commit:

- Moves /pricing, /privacy, /terms, /billing-refunds-test, and the four
  buyer landings to /marketing/* (real prefix, no route group)
- Adds /marketing as a hub page that lists the four buyer landings
- Keeps / (home) at the root via a thin (marketing)/layout.tsx wrapper;
  documents in a comment that adding new routes to that group is an
  anti-pattern
- Extracts the marketing nav + footer into a shared MarketingShell so the
  two layout files do not drift
- Adds Solutions column to the footer linking all four buyer landings, plus
  a Solutions link in the top nav pointing to /marketing
- Adds 308 redirects in next.config.js for every legacy URL so external
  links keep working
- Updates src/proxy.ts to whitelist /marketing as a public-path prefix
- Updates sitemap.ts with the new canonical URLs including the buyer pages
  (which were missing from the sitemap entirely before)
- Updates every internal Link href, every revalidatePath call, and every
  e2e test goto to the new canonical URLs

Verification:
- pnpm run typecheck: clean
- pnpm run test:unit: 1046 tests passing (including new proxy assertions
  for /marketing/* public path)
- curl http://localhost/pricing -> 308 to /marketing/pricing (verified)
- curl http://localhost/vp-engineering -> 308 to /marketing/vp-engineering
  (verified)

Refs CHAOS-1647
