+++ src/app/marketing/cto-architecture/page.tsx
@@ -2,89 +2,87 @@ import type { Metadata } from "next";
 import Link from "next/link";
 
 export const metadata: Metadata = {
-  title: "For Platform / DevEx — Full Chaos Dev Health",
-  description: "Find systemic friction across repos, CI, reviews, and deployments.",
+  title: "For CTO / Architecture — Full Chaos Dev Health",
+  description: "See where change pressure is compounding architectural risk.",
 };
 
-const SURFACES: Array<{ label: string; title: string; description: string; href: string; icon: React.ReactElement; comingSoon?: boolean }> = [
+const SURFACES = [
   {
-    label: "CI/CD + TestOps",
-    title: "CI/CD + TestOps",
-    description: "Find systemic friction across repos, CI, reviews, and deployments.",
-    href: "/testops",
+    title: "Compounding Risk",
+    href: "/risk/compounding",
+    comingSoon: true,
+    description: "Track where high churn meets high complexity to identify system fragility.",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <path d="M14 14.76V3.5a2.5 2.5 0 0 0-5 0v11.26a4.5 4.5 0 1 0 5 0z" />
+        <path d="M10.29 3.86L1.82 18a2 2 0 0 0 1.71 3h16.94a2 2 0 0 0 1.71-3L13.71 3.86a2 2 0 0 0-3.42 0z" />
+        <line x1="12" y1="9" x2="12" y2="13" />
+        <line x1="12" y1="17" x2="12.01" y2="17" />
       </svg>
     ),
   },
   {
-    label: "Review Load",
-    title: "Review Load",
-    description: "Identify bottlenecks in PR reviews and systemic latency in the code review process.",
-    href: "/ai/review-load",
+    title: "Complexity trends",
+    href: "/complexity",
+    comingSoon: true,
+    description: "Monitor code complexity over time to prevent unmanageable technical debt.",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <path d="M22 12h-2.48a2 2 0 0 0-1.93 1.46l-2.35 8.36a.25.25 0 0 1-.48 0L9.24 2.18a.25.25 0 0 0-.48 0l-2.35 8.36A2 2 0 0 1 4.49 12H2" />
+        <polyline points="22 12 18 12 15 21 9 3 6 12 2 12" />
       </svg>
     ),
   },
   {
-    label: "Work Graph",
-    title: "Work Graph",
-    description: "Map the dependencies and connections between teams, repos, and issues.",
-    href: "/work",
-    icon: (
-      <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <circle cx="18" cy="5" r="3" />
-        <circle cx="6" cy="12" r="3" />
-        <circle cx="18" cy="19" r="3" />
-        <line x1="8.59" x2="15.42" y1="13.51" y2="17.49" />
-        <line x1="15.41" x2="8.59" y1="6.51" y2="10.49" />
-      </svg>
-    ),
-  },
-  {
-    label: "Hotspots",
-    title: "Hotspots",
-    description: "Discover complex files with high churn that slow down feature delivery.",
+    title: "Ownership concentration",
     href: "/code",
+    description: "Identify key person dependencies and knowledge silos before they become bottlenecks.",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <path d="M12 2v20" />
-        <path d="M17 5H9.5a3.5 3.5 0 0 0 0 7h5a3.5 3.5 0 0 1 0 7H6" />
+        <path d="M17 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2" />
+        <circle cx="9" cy="7" r="4" />
+        <path d="M23 21v-2a4 4 0 0 0-3-3.87" />
+        <path d="M16 3.13a4 4 0 0 1 0 7.75" />
       </svg>
     ),
   },
   {
-    label: "Data Health",
-    title: "Connector / Data Health",
-    description: "Monitor the reliability of data from Jira, GitHub, and GitLab connectors.",
-    href: "/data-health",
+    title: "Hotspots",
+    href: "/code",
+    description: "Pinpoint the areas of your codebase consuming the most effort and generating the most bugs.",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <ellipse cx="12" cy="5" rx="9" ry="3" />
-        <path d="M3 5V19A9 3 0 0 0 21 19V5" />
-        <path d="M3 12A9 3 0 0 0 21 12" />
+        <path d="M8.5 14.5A2.5 2.5 0 0 0 11 12c0-1.38-.5-2-1-3-1.072-2.143-.224-4.054 2-6 .5 2.5 2 4.9 4 6.5 2 1.6 3 3.5 3 5.5a7 7 0 1 1-14 0c0-1.153.433-2.294 1-3a2.5 2.5 0 0 0 2.5 2.5z" />
+      </svg>
+    ),
+  },
+  {
+    title: "Incident correlation",
+    href: "/incident-correlation",
+    comingSoon: true,
+    description: "Connect system outages to recent deployment activity to speed up root cause analysis.",
+    icon: (
+      <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+        <path d="M18.36 6.64a9 9 0 1 1-12.73 0" />
+        <line x1="12" y1="2" x2="12" y2="12" />
       </svg>
     ),
   },
 ];
 
-export default function PlatformDevexPage() {
+export default function CTOArchitecturePage() {
   return (
     <>
       {/* Hero */}
       <section className="mx-auto max-w-7xl px-6 pb-20 pt-16 sm:pt-24">
         <div className="mx-auto max-w-3xl text-center">
           <p className="text-xs uppercase tracking-[0.15em] text-(--ink-muted)">
-            For Platform / DevEx
+            For CTO / Architecture
           </p>
           <h1 className="mt-6 font-(--font-display) text-4xl leading-tight sm:text-5xl lg:text-6xl">
-            Find systemic friction across repos, CI, reviews, and deployments
+            See where change pressure is compounding{" "}
+            <span className="text-(--accent)">architectural risk</span>
           </h1>
           <p className="mx-auto mt-6 max-w-2xl text-lg text-(--ink-muted)">
-            Identify systemic blockers, reduce coordination debt, and monitor the health of your delivery pipeline. Optimize the developer experience across your entire organization.
+            Align engineering effort with architectural reality. Identify systemic bottlenecks, key person dependencies, and fragile codebases before they compromise your delivery velocity.
           </p>
           <div className="mt-10 flex flex-wrap items-center justify-center gap-4">
             <Link
@@ -94,7 +92,7 @@ export default function PlatformDevexPage() {
               Start for free
             </Link>
             <Link
-              href="/pricing"
+              href="/marketing/pricing"
               className="rounded-full border border-(--card-stroke) bg-(--card-70) px-8 py-3 text-sm font-medium transition hover:border-foreground/30"
             >
               See pricing
@@ -103,20 +101,24 @@ export default function PlatformDevexPage() {
         </div>
       </section>
 
-      {/* Primary Surfaces */}
+      {/* Primary surfaces */}
       <section className="mx-auto max-w-7xl px-6 pb-24">
-        <div className="grid gap-5 sm:grid-cols-2 lg:grid-cols-3">
+        <div className="text-center">
+          <p className="text-xs uppercase tracking-[0.15em] text-(--ink-muted)">
+            Primary surfaces
+          </p>
+          <h2 className="mt-4 font-(--font-display) text-3xl sm:text-4xl">
+            Insights mapped to architectural health
+          </h2>
+        </div>
+
+        <div className="mt-14 grid gap-5 sm:grid-cols-2 lg:grid-cols-3">
           {SURFACES.map((surface) => {
             const cardInner = (
               <>
                 <div className="flex items-center justify-between gap-3">
-                  <div className="flex items-center gap-3">
-                    <div className="flex size-9 items-center justify-center rounded-xl bg-(--accent)/10 text-(--accent)">
-                      {surface.icon}
-                    </div>
-                    <p className="text-[10px] uppercase tracking-[0.2em] text-(--ink-muted)">
-                      {surface.label}
-                    </p>
+                  <div className="flex size-9 items-center justify-center rounded-xl bg-(--accent)/10 text-(--accent)">
+                    {surface.icon}
                   </div>
                   {surface.comingSoon && (
                     <span className="rounded-full border border-(--card-stroke) bg-(--card) px-2 py-0.5 text-[9px] font-semibold uppercase tracking-[0.2em] text-(--ink-muted)">
@@ -124,7 +126,7 @@ export default function PlatformDevexPage() {
                     </span>
                   )}
                 </div>
-                <h3 className="mt-4 font-(--font-display) text-lg">
+                <h3 className="mt-4 font-(--font-display) text-lg group-hover:text-(--accent) transition-colors">
                   {surface.title}
                 </h3>
                 <p className="mt-2 text-sm leading-relaxed text-(--ink-muted)">
@@ -135,7 +137,7 @@ export default function PlatformDevexPage() {
             if (surface.comingSoon) {
               return (
                 <div
-                  key={surface.label}
+                  key={surface.title}
                   className="block rounded-3xl border border-(--card-stroke) bg-(--card-80) p-6 opacity-90"
                   aria-disabled="true"
                 >
@@ -145,7 +147,7 @@ export default function PlatformDevexPage() {
             }
             return (
               <Link
-                key={surface.label}
+                key={surface.title}
                 href={surface.href}
                 className="group block rounded-3xl border border-(--card-stroke) bg-(--card-80) p-6 transition hover:-translate-y-1"
               >
@@ -175,15 +177,40 @@ export default function PlatformDevexPage() {
 
           <ul className="mx-auto mt-10 grid max-w-4xl gap-3 sm:grid-cols-2">
             {[
-              { id: "ns-no-leaderboards", title: "No leaderboards", body: "We do not rank people against each other. Ever." },
-              { id: "ns-team-first", title: "Team and repo first", body: "Aggregation defaults to systems, not individuals." },
-              { id: "ns-reflection-only", title: "Individual views are reflection-only", body: "Single-person surfaces exist for self-reflection and coaching — never peer comparison." },
-              { id: "ns-trends-over-absolutes", title: "Trends over absolutes", body: "Direction matters more than point values. We benchmark you against your own baseline." },
-              { id: "ns-evidence-over-scores", title: "Evidence over scores", body: "Every signal links back to the PRs, reviews, commits, and incidents behind it." },
+              {
+                id: "ns-no-leaderboards",
+                title: "No leaderboards",
+                body: "We do not rank people against each other. Ever.",
+              },
+              {
+                id: "ns-team-first",
+                title: "Team and repo first",
+                body: "Aggregation defaults to systems, not individuals.",
+              },
+              {
+                id: "ns-reflection-only",
+                title: "Individual views are reflection-only",
+                body: "Single-person surfaces exist for self-reflection and coaching — never peer comparison.",
+              },
+              {
+                id: "ns-trends-over-absolutes",
+                title: "Trends over absolutes",
+                body: "Direction matters more than point values. We benchmark you against your own baseline.",
+              },
+              {
+                id: "ns-evidence-over-scores",
+                title: "Evidence over scores",
+                body: "Every signal links back to the PRs, reviews, commits, and incidents behind it.",
+              },
             ].map((pillar) => (
-              <li key={pillar.id} className="rounded-2xl border border-(--card-stroke) bg-(--card) p-5">
+              <li
+                key={pillar.id}
+                className="rounded-2xl border border-(--card-stroke) bg-(--card) p-5"
+              >
                 <p className="font-(--font-display) text-base">{pillar.title}</p>
-                <p className="mt-2 text-sm leading-relaxed text-(--ink-muted)">{pillar.body}</p>
+                <p className="mt-2 text-sm leading-relaxed text-(--ink-muted)">
+                  {pillar.body}
+                </p>
               </li>
             ))}
           </ul>
@@ -194,7 +221,7 @@ export default function PlatformDevexPage() {
         </div>
       </section>
 
-      {/* Final CTA */}
+      {/* CTA */}
       <section className="mx-auto max-w-7xl px-6 pb-24">
         <div className="rounded-3xl border border-(--card-stroke) bg-(--card-80) p-8 text-center sm:p-12">
           <p className="text-xs uppercase tracking-[0.15em] text-(--ink-muted)">
@@ -215,14 +242,12 @@ export default function PlatformDevexPage() {
             >
               Get started free
             </Link>
-            <a
+            <Link
               href="https://github.com/full-chaos/dev-health-ops"
-              target="_blank"
-              rel="noopener"
               className="rounded-full border border-(--card-stroke) bg-(--card-70) px-8 py-3 text-sm font-medium transition hover:border-foreground/30"
             >
               Star on GitHub
-            </a>
+            </Link>
           </div>
         </div>
       </section>
