+++ src/app/marketing/platform-devex/page.tsx
@@ -1,92 +1,90 @@
-import { Metadata } from "next";
+import type { Metadata } from "next";
 import Link from "next/link";
 
 export const metadata: Metadata = {
-  title: "For VP Engineering — Full Chaos Dev Health",
-  description: "Know where delivery is constrained before it becomes a miss.",
+  title: "For Platform / DevEx — Full Chaos Dev Health",
+  description: "Find systemic friction across repos, CI, reviews, and deployments.",
 };
 
-const SURFACES = [
+const SURFACES: Array<{ label: string; title: string; description: string; href: string; icon: React.ReactElement; comingSoon?: boolean }> = [
   {
-    label: "Operating Review",
-    title: "Engineering Operating Review",
-    description: "Align engineering outputs with business outcomes. See organizational throughput and delivery pacing at a glance.",
-    href: "/operating-review",
+    label: "CI/CD + TestOps",
+    title: "CI/CD + TestOps",
+    description: "Find systemic friction across repos, CI, reviews, and deployments.",
+    href: "/testops",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <path d="M3 3v18h18" />
-        <path d="m19 9-5 5-4-4-3 3" />
+        <path d="M14 14.76V3.5a2.5 2.5 0 0 0-5 0v11.26a4.5 4.5 0 1 0 5 0z" />
       </svg>
     ),
   },
   {
-    label: "Capacity Forecast",
-    title: "Predictable Capacity",
-    description: "Forecast delivery constraints based on historical throughput, cycle time trends, and WIP saturation.",
-    href: "/capacity-planning",
+    label: "Review Load",
+    title: "Review Load",
+    description: "Identify bottlenecks in PR reviews and systemic latency in the code review process.",
+    href: "/ai/review-load",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <circle cx="12" cy="12" r="10" />
-        <path d="M12 6v6l4 2" />
+        <path d="M22 12h-2.48a2 2 0 0 0-1.93 1.46l-2.35 8.36a.25.25 0 0 1-.48 0L9.24 2.18a.25.25 0 0 0-.48 0l-2.35 8.36A2 2 0 0 1 4.49 12H2" />
       </svg>
     ),
   },
   {
-    label: "Bottlenecks",
-    title: "Delivery Bottleneck Summary",
-    description: "Spot where work is piling up. Identify high review latency and coordination debt before it impacts ship dates.",
-    href: "/bottleneck",
-    comingSoon: true,
+    label: "Work Graph",
+    title: "Work Graph",
+    description: "Map the dependencies and connections between teams, repos, and issues.",
+    href: "/work",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <path d="M5 22h14" />
-        <path d="M5 2h14" />
-        <path d="M17 22V2" />
-        <path d="M7 22V2" />
+        <circle cx="18" cy="5" r="3" />
+        <circle cx="6" cy="12" r="3" />
+        <circle cx="18" cy="19" r="3" />
+        <line x1="8.59" x2="15.42" y1="13.51" y2="17.49" />
+        <line x1="15.41" x2="8.59" y1="6.51" y2="10.49" />
       </svg>
     ),
   },
   {
-    label: "Investment",
-    title: "Investment Allocation",
-    description: "See the exact mix of feature delivery, operational support, maintenance, and technical debt across the organization.",
-    href: "/investment",
+    label: "Hotspots",
+    title: "Hotspots",
+    description: "Discover complex files with high churn that slow down feature delivery.",
+    href: "/code",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <path d="M21.21 15.89A10 10 0 1 1 8 2.83" />
-        <path d="M22 12A10 10 0 0 0 12 2v10z" />
+        <path d="M12 2v20" />
+        <path d="M17 5H9.5a3.5 3.5 0 0 0 0 7h5a3.5 3.5 0 0 1 0 7H6" />
       </svg>
     ),
   },
   {
-    label: "Reliability",
-    title: "Reliability & Incidents",
-    description: "Connect incident volume to delivery velocity. Understand the cost of poor quality on feature throughput.",
-    href: "/quality",
+    label: "Data Health",
+    title: "Connector / Data Health",
+    description: "Monitor the reliability of data from Jira, GitHub, and GitLab connectors.",
+    href: "/data-health",
     icon: (
       <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-        <path d="m3 11 18-5v12L3 13v-2z" />
-        <path d="M11.6 16.8a3 3 0 1 1-5.8-1.6" />
+        <ellipse cx="12" cy="5" rx="9" ry="3" />
+        <path d="M3 5V19A9 3 0 0 0 21 19V5" />
+        <path d="M3 12A9 3 0 0 0 21 12" />
       </svg>
     ),
   },
 ];
 
-
-export default function VPEngineeringPage() {
+export default function PlatformDevexPage() {
   return (
     <>
       {/* Hero */}
       <section className="mx-auto max-w-7xl px-6 pb-20 pt-16 sm:pt-24">
         <div className="mx-auto max-w-3xl text-center">
           <p className="text-xs uppercase tracking-[0.15em] text-(--ink-muted)">
-            For VP Engineering
+            For Platform / DevEx
           </p>
           <h1 className="mt-6 font-(--font-display) text-4xl leading-tight sm:text-5xl lg:text-6xl">
-            Where is delivery <span className="text-(--accent)">constrained</span>?
+            Find systemic friction across repos, CI, reviews, and deployments
           </h1>
           <p className="mx-auto mt-6 max-w-2xl text-lg text-(--ink-muted)">
-            Know where delivery is constrained before it becomes a miss.
+            Identify systemic blockers, reduce coordination debt, and monitor the health of your delivery pipeline. Optimize the developer experience across your entire organization.
           </p>
           <div className="mt-10 flex flex-wrap items-center justify-center gap-4">
             <Link
@@ -96,7 +94,7 @@ export default function VPEngineeringPage() {
               Start for free
             </Link>
             <Link
-              href="/pricing"
+              href="/marketing/pricing"
               className="rounded-full border border-(--card-stroke) bg-(--card-70) px-8 py-3 text-sm font-medium transition hover:border-foreground/30"
             >
               See pricing
@@ -107,15 +105,7 @@ export default function VPEngineeringPage() {
 
       {/* Primary Surfaces */}
       <section className="mx-auto max-w-7xl px-6 pb-24">
-        <div className="text-center">
-          <p className="text-xs uppercase tracking-[0.15em] text-(--ink-muted)">
-            Primary Surfaces
-          </p>
-          <h2 className="mt-4 font-(--font-display) text-3xl sm:text-4xl">
-            Map constraints to reality
-          </h2>
-        </div>
-        <div className="mt-14 grid gap-5 sm:grid-cols-2 lg:grid-cols-3">
+        <div className="grid gap-5 sm:grid-cols-2 lg:grid-cols-3">
           {SURFACES.map((surface) => {
             const cardInner = (
               <>
@@ -204,7 +194,7 @@ export default function VPEngineeringPage() {
         </div>
       </section>
 
-      {/* CTA */}
+      {/* Final CTA */}
       <section className="mx-auto max-w-7xl px-6 pb-24">
         <div className="rounded-3xl border border-(--card-stroke) bg-(--card-80) p-8 text-center sm:p-12">
           <p className="text-xs uppercase tracking-[0.15em] text-(--ink-muted)">
@@ -214,7 +204,9 @@ export default function VPEngineeringPage() {
             Built in the open, for everyone
           </h2>
           <p className="mx-auto mt-4 max-w-xl text-sm text-(--ink-muted)">
-            Full Chaos Dev Health is fully open source. Deploy it on your infrastructure, audit every metric computation, and contribute to the roadmap. Accessibility over extraction.
+            Full Chaos Dev Health is fully open source. Deploy it on your infrastructure,
+            audit every metric computation, and contribute to the roadmap.
+            Accessibility over extraction.
           </p>
           <div className="mt-8 flex flex-wrap items-center justify-center gap-4">
             <Link
@@ -223,14 +215,14 @@ export default function VPEngineeringPage() {
             >
               Get started free
             </Link>
-            <Link
+            <a
               href="https://github.com/full-chaos/dev-health-ops"
               target="_blank"
-              rel="noopener noreferrer"
+              rel="noopener"
               className="rounded-full border border-(--card-stroke) bg-(--card-70) px-8 py-3 text-sm font-medium transition hover:border-foreground/30"
             >
               Star on GitHub
-            </Link>
+            </a>
           </div>
         </div>
       </section>
