+++ src/app/marketing/terms/page.tsx
